Optimise status code conditions
The current conditions to match which level of log to use are more expensive that they should be. Instead of building a range object and do a lookup in it, which -in the worst case- can go through 100 of elements, it's 2x faster to compare with the 2 integers of the bound:

```
>>> import timeit
>>> timeit.timeit("x in range(400, 500)", setup="from random import randint;x = randint(100, 1000)")
0.08440266699471977
>>> timeit.timeit("400 <= x < 500", setup="from random import randint;x = randint(100, 1000)")
0.0354189580102684
```

As this middleware runs on every request, and info logs go through the if + elif condition, I think this small optimisation, while not life changing, is nice to have.